Add default particles
